refactor: 상품 가격의 타입 변경
- Integer -> Long

+++ src/main/java/gift/dto/CreateProductResponse.java
@@ -2,7 +2,7 @@ package gift.dto;
 
 import gift.domain.Product;
 
-public record CreateProductResponse(Long id, String name, Integer price, String imageUrl) {
+public record CreateProductResponse(Long id, String name, Long price, String imageUrl) {
 
     public static CreateProductResponse from(Product product) {
         return new CreateProductResponse(
