+++ src/main/java/gift/dto/ProductResponse.java
@@ -2,7 +2,7 @@ package gift.dto;
 
 import gift.domain.Product;
 
-public record ProductResponse(Long id, String name, Integer price, String imageUrl) {
+public record ProductResponse(Long id, String name, Long price, String imageUrl) {
 
     public static ProductResponse from(Product product) {
         return new ProductResponse(
